# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Documentation API for Coffeshop Senja Kopi Kiri",
        "title": "Coffeeshop Senja Kopi Kiri",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8011",
    "basePath": "/",
    "paths": {
...
        "/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets a list of products in the cart of the logged in user.",
                "tags": [
                    "Cart"
                ],
                "summary": "Get cart products",
                "responses": {
                    "200": {
                        "description": "Cart data retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSucces"
                        }
                    },
                    "401": {
                        "description": "User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Failed get data carts",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adding products to the cart of the logged in user",
                "tags": [
                    "Cart"
                ],
                "summary": "Add product to cart",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product added to cart successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSucces"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON, validation error, out of stock, or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: user not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/cart/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete specific cart item based on cart ID and user ID from JWT token",
                "tags": [
                    "Cart"
                ],
                "summary": "Delete cart item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete cart successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSucces"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid cart ID",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Cart not found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to delete cart",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/transactions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Performs a transaction for the authenticated user. Includes validation and business logic checks.",
                "tags": [
                    "Transactions"
                ],
                "summary": "Process a transaction",
                "parameters": [
                    {
                        "description": "Transaction Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.RequestTransactions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction completed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSucces"
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid JSON format",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: user not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CartItemRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer",
                    "maximum": 3
                },
                "variant": {
                    "type": "integer",
                    "maximum": 2
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.ResponseSucces": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.RequestTransactions": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "jakarta"
                },
                "email": {
                    "type": "string",
                    "example": "youremail@gmail.com"
                },
                "fullname": {
                    "type": "string",
                    "example": "fullname"
                },
                "id_delivery": {
                    "type": "integer"
                },
                "id_paymentMethod": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string",
                    "example": "081234567890"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}